+++ typescript/zetachain/zetacore/pkg/chains/chains_pb.d.ts
@@ -148,6 +148,11 @@ export declare enum ChainName {
    * @generated from enum value: solana_localnet = 23;
    */
   solana_localnet = 23,
+
+  /**
+   * @generated from enum value: btc_signet_testnet = 24;
+   */
+  btc_signet_testnet = 24,
 }
 
 /**