+++ zetaclient/orchestrator/bootstap_test.go
@@ -49,7 +49,7 @@ func TestCreateSignerMap(t *testing.T) {
 			Endpoint: testutils.MockEVMRPCEndpoint,
 		}
 
-		cfg.BitcoinConfig = btcConfig
+		cfg.BTCChainConfigs[chains.BitcoinMainnet.ChainId] = btcConfig
 
 		// Given AppContext
 		app := zctx.New(cfg, nil, log)
@@ -227,7 +227,7 @@ func TestCreateChainObserverMap(t *testing.T) {
 			Endpoint: evmServer.Endpoint,
 		}
 
-		cfg.BitcoinConfig = btcConfig
+		cfg.BTCChainConfigs[chains.BitcoinMainnet.ChainId] = btcConfig
 		cfg.SolanaConfig = solConfig
 
 		// Given AppContext